Fix for when Go is used from source without release tag
